requirements.txt
wheel 
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt
Pillow

psycopg2-binary
python-decouple
python-slugify

twilio
celery
redis

oauth2_provider
social-auth-app-django
drf_social_oauth2
django-cors-headers